todo_api_checks.c 2

+++ sockets/todo_api_checks.c
@@ -118,11 +118,11 @@ return NULL;
 }
 snprintf(response, response_size + 1,
 "HTTP/1.1 201 Created\r\n"
-"Content-Length: %zu\r\n"
+"Content-Length: %lu\r\n"
 "Content-Type: application/json\r\n"
 "\r\n"
 "%s",
-body_size, json_body);
+(unsigned long)body_size, json_body);
 
 free(json_body);
 return response;
